requirements.txt
mcp
azure-ai-agents
azure-identity
python-dotenv
msal
aiohttp
